+++ Claudius/Vector2.cpp
@@ -28,12 +28,12 @@ Vector2 Vector2::operator/(const Vector2& rhs) const
 	return Vector2(x / rhs.x, y / rhs.y);
 }
 
-Vector2 Vector2::operator*(const float rhs) const
+Vector2 Vector2::operator*(const int rhs) const
 {
 	return Vector2(x * rhs, y * rhs);
 }
 
-Vector2 Vector2::operator/(const float rhs) const
+Vector2 Vector2::operator/(const int rhs) const
 {
 	return Vector2(x / rhs, y / rhs);
 }
