+++ Claudius/Vector2.h
@@ -2,16 +2,16 @@
 
 struct Vector2
 {
-	float x {0}; 
-	float y{ 0 };
+	int x {0}; 
+	int y{ 0 };
 
 	Vector2& operator=(const Vector2& rhs);				 
 	Vector2 operator+(const Vector2& rhs) const;		 
 	Vector2 operator-(const Vector2& rhs) const;		 
 	Vector2 operator*(const Vector2& rhs) const;	
 	Vector2 operator/(const Vector2& rhs) const;
-	Vector2 operator*(const float rhs) const;			 
-	Vector2 operator/(const float rhs) const;		
+	Vector2 operator*(const int rhs) const;			 
+	Vector2 operator/(const int rhs) const;		
 	bool operator==(const Vector2& rhs) const;
 
 };
